reviewer/requirements.txt
google-generativeai==0.7.2
PyGithub==2.3.0
requests==2.32.0
python-dotenv==1.0.1
